fix: Negative residual in version generation leads to endless loop during pre-aggregation refresh
